Further splitting up serializer: - Separating out logging (and tryQuery which only exists to guard and log dangerous accesses) - Separating out all module-related logic
